Mention message level in output when it should be
The intention of BriefFormatter was to suppress the level prefix for
informal messages but show it for others.

This slipped unnoticed because most (if not all) error-level messages
are actually printed by other means and the warning-level messages are
too rare.

This fixes commit 93d9439c0c8642e00df2dd378f9df261c5bab02a (Make logging
briefer unless debugging is on)